sticker/requirements_auth.txt
requests==2.31.0
telethon==1.36.0
PyJWT==2.10.1
aiohttp==3.9.1
python-dotenv==1.2.1

